修复Xcode 4.5下出现的几个问题

+++ DynamicLyrics/LyricX/Albumfiller.m
@@ -74,7 +74,7 @@
     [DouBan SearchArtwork:[[iTunes currentTrack] album]];
     
     SBElementArray* theArtworks = [[iTunes currentTrack] artworks];
-    int totalArtworkCount = [theArtworks count];
+    unsigned long totalArtworkCount = [theArtworks count];
     if (totalArtworkCount > 0) {
         iTunesArtwork *thisArtwork = [theArtworks objectAtIndex:0];
